Added isi fragment to page default
